The API is functional.
The API is functional, commented and formatted. At the moment the relationships are all over the place, but they work. IF...IF I have some time left I will take another shot at this since it leaves a bat taste in my mouth leaving the code like this. Nonetheless It does what it's tasked with.

- Dudu out.